FP MealMap Website/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2024.10
cryptography==44.0.1
daphne==4.1.2
Django==4.2.16
django-bootstrap4==24.4
django-ckeditor==6.7.2
django-countries==7.6.1
django-crispy-forms==2.3
django-environ==0.11.2
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.9
environ==1.0
factory_boy==3.3.1
Faker==30.8.1
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
phonenumbers==8.13.53
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
sendgrid==6.11.0
service-identity==24.2.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tomli==2.2.1
twilio==9.3.7
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
yarl==1.18.0
zope.interface==7.2
